# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SIMMAS API",
        "description": "Sistem Informasi Manajemen Magang Siswa (SIMMAS). REST API dengan autentikasi Sanctum. Token didapat dari endpoint `POST /api/login`. Semua endpoint selain `login` dan `public/*` memerlukan header `Authorization: Bearer <token>`.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Server backend SIMMAS (Laravel)"
        }
    ],
    "paths": {
...
        "/api/admin/penempatan": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Daftar penempatan magang",
                "description": "Semua penempatan beserta relasi siswa, tempat magang, dan guru pembimbing.",
                "operationId": "581687b8140af21bd0fa2ffc9c3b5c93",
                "responses": {
                    "200": {
                        "description": "Daftar penempatan.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PenempatanMagang"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Buat penempatan magang",
                "description": "Menempatkan siswa ke DUDI + guru pembimbing, sekaligus mengubah status siswa menjadi sedang_magang. Kuota DUDI harus tersedia.",
                "operationId": "ea6cf6084a7a0d27e91e3c96d16a02ed",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "siswa_id",
                                    "tempat_magang_id",
                                    "guru_id",
                                    "tanggal_mulai",
                                    "tanggal_selesai"
                                ],
                                "properties": {
                                    "siswa_id": {
                                        "type": "integer"
                                    },
                                    "tempat_magang_id": {
                                        "type": "integer"
                                    },
                                    "guru_id": {
                                        "type": "integer"
                                    },
                                    "tanggal_mulai": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "tanggal_selesai": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Penempatan berhasil dibuat.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PenempatanMagang"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal atau kuota DUDI habis."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/penempatan/{id}": {
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "Perbarui penempatan magang",
                "operationId": "08a61195cb2f72ae9544e1c9895bbd20",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tempat_magang_id",
                                    "guru_id",
                                    "tanggal_mulai",
                                    "tanggal_selesai"
                                ],
                                "properties": {
                                    "tempat_magang_id": {
                                        "type": "integer"
                                    },
                                    "guru_id": {
                                        "type": "integer"
                                    },
                                    "tanggal_mulai": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "tanggal_selesai": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Penempatan berhasil diperbarui.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PenempatanMagang"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/penempatan/{id}/sahkan": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Sahkan penempatan magang",
                "description": "Mengesahkan penempatan dan menandai siswa sebagai sedang_magang.",
                "operationId": "1dd91f184e3c52e6c0b1f8d0c1864107",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Penempatan disahkan.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PenempatanMagang"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/penempatan/{id}/batalkan": {
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Batalkan penempatan magang",
                "description": "Menghapus penempatan dan mengembalikan status siswa menjadi belum_magang.",
                "operationId": "c61c174e7a9735f94437df8f9f536758",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Penempatan dibatalkan."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/admin/settings": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Lihat pengaturan sistem",
                "description": "Mengembalikan identitas aplikasi, konten halaman depan, dan profil sekolah.",
                "operationId": "b8df30762e1031b1797e3181636b893d",
                "responses": {
                    "200": {
                        "description": "Pengaturan sistem."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "Simpan pengaturan sistem",
                "description": "Menyimpan perubahan identitas aplikasi, halaman depan, dan data sekolah. Tercatat di activity log.",
                "operationId": "1aa85c87d56da8647682bd8feab7f1b8",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "identitas": {
                                        "type": "object"
                                    },
                                    "halaman_depan": {
                                        "type": "object"
                                    },
                                    "sekolah": {
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pengaturan berhasil disimpan."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PenempatanMagang": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "siswa_id": {
                        "type": "integer"
                    },
                    "tempat_magang_id": {
                        "type": "integer"
                    },
                    "guru_id": {
                        "type": "integer"
                    },
                    "tanggal_mulai": {
                        "type": "string",
                        "format": "date"
                    },
                    "tanggal_selesai": {
                        "type": "string",
                        "format": "date"
                    },
                    "nilai_akhir": {
                        "type": "integer",
                        "nullable": true
                    },
                    "status_pengesahan": {
                        "type": "string",
                        "enum": [
                            "belum_disahkan",
                            "disahkan",
                            "lulus_magang"
                        ]
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Masukkan token dengan format: Bearer <token> (didapat dari endpoint POST /api/login).",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}